* moar warning nerfs. * killbyaccount/player/ip functions will now return a "nobody online with account/ip/name [%s]" if no one is found online, rather then saying nothing at all.
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@4520 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/ascent-world/Stats.cpp
@@ -643,7 +643,7 @@ uint32 CalculateDamage( Unit* pAttacker, Unit* pVictim, uint32 weapon_damage_typ
 	{
 		if( pAttacker->IsPlayer() && ((Player*)pAttacker)->m_outStealthDamageBonusTimer )
 		{
-			if( UNIXTIME >= ((Player*)pAttacker)->m_outStealthDamageBonusTimer )
+			if( (uint32)UNIXTIME >= ((Player*)pAttacker)->m_outStealthDamageBonusTimer )
 				((Player*)pAttacker)->m_outStealthDamageBonusTimer = 0;
 			else
 				result *= ((float(((Player*)pAttacker)->m_outStealthDamageBonusPct) / 100.0f) + 1.0f);
